Add load handler
Add the (untested) load handler. All of the load requests will be
directed to the load handler. There are two decisions made here which
may change later.

* Right now CPU usage percentage is measured by calling the C `clock`
  method at the start and the end of some computationally intensive
  method, and determining the average amount of CPU consumed per second
  during that time. This seems like a pretty good method, but I'll have
  to check if this corresponds with how k8s measures CPU, and also it
  would be nice to not be running C code from golang.
* The cost intensive task right now is creating a large number of cost
  intensive hashed passwords. It may be necessary to do some testing to
  see if this task occupies an appropriate amount of CPU.

+++ evaluation/test-server/pkg/database/influx.go
@@ -44,7 +44,7 @@ func createClient() (client.Client, error) {
 
 // WriteMetrics is called from a handler to write the eru and qos metrics to the
 // influxdb database.
-func WriteMetrics(eru, qos float32) error {
+func WriteMetrics(eru, qos float64) error {
 	influxClient, err := createClient()
 
 	if err != nil {
